@@ -34,19 +34,23 @@
     "@emotion/react": "^11.10.5",
     "@emotion/styled": "^11.10.5",
     "@mdx-js/react": "^2.1.5",
-    "gatsby": "^5.0.0",
+    "gatsby": "^5.2.0",
     "gatsby-plugin-alias-imports": "^1.0.5",
-    "gatsby-plugin-emotion": "^8.1.0",
+    "gatsby-plugin-emotion": "^8.2.0",
     "gatsby-plugin-feed": "^5.2.0",
-    "gatsby-plugin-image": "^3.0.0",
-    "gatsby-plugin-manifest": "^5.0.0",
-    "gatsby-plugin-mdx": "^5.0.0",
-    "gatsby-plugin-offline": "^6.0.0",
+    "gatsby-plugin-image": "^3.2.0",
+    "gatsby-plugin-manifest": "^5.2.0",
+    "gatsby-plugin-offline": "^6.2.0",
     "gatsby-plugin-robots-txt": "^1.8.0",
-    "gatsby-plugin-sharp": "^5.0.0",
-    "gatsby-plugin-sitemap": "^6.0.0",
-    "gatsby-source-filesystem": "^5.0.0",
-    "gatsby-transformer-sharp": "^5.0.0",
+    "gatsby-plugin-sharp": "^5.2.0",
+    "gatsby-plugin-sitemap": "^6.2.0",
+    "gatsby-remark-autolink-headers": "^6.2.0",
+    "gatsby-remark-emoji": "^0.0.3",
+    "gatsby-remark-prismjs": "^7.2.0",
+    "gatsby-source-filesystem": "^5.2.0",
+    "gatsby-transformer-remark": "^6.2.0",
+    "gatsby-transformer-sharp": "^5.2.0",
+    "prismjs": "^1.29.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "twin.macro": "^3.0.1"
@@ -55,9 +59,9 @@
     "@graphql-eslint/eslint-plugin": "^3.13.1",
     "@types/node": "^18.11.9",
     "@types/react": "^18.0.25",
-    "@types/react-dom": "^18.0.8",
-    "@typescript-eslint/eslint-plugin": "^5.42.1",
-    "@typescript-eslint/parser": "^5.42.1",
+    "@types/react-dom": "^18.0.9",
+    "@typescript-eslint/eslint-plugin": "^5.44.0",
+    "@typescript-eslint/parser": "^5.44.0",
     "autoprefixer": "^10.4.13",
     "eslint": "^8.28.0",
     "eslint-config-prettier": "^8.5.0",
@@ -66,13 +70,13 @@
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-jsx-a11y": "^6.6.1",
     "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-react": "^7.31.10",
+    "eslint-plugin-react": "^7.31.11",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-unused-imports": "^2.0.0",
-    "gatsby-plugin-postcss": "^6.0.0",
+    "gatsby-plugin-postcss": "^6.2.0",
     "postcss": "^8.4.19",
     "prettier": "2.7.1",
     "tailwindcss": "^3.2.4",
-    "typescript": "^4.8.4"
+    "typescript": "^4.9.3"
   }
 }
